requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.2.26
langchain-openai
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-checkpoint-postgres
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
MarkupSafe==3.0.3
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
psycopg[binary]
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.43.0
xxhash==3.6.0
zstandard==0.25.0
